add userId in DeleteMealTicketRemovalCommand and handler

+++ server/src/Infrastructure/HumanResource/MealTicket/Repository/MealTicketRemovalRepository.ts
@@ -68,11 +68,12 @@ export class MealTicketRemovalRepository
       .getRawOne();
   }
 
-  public async deleteOne(id: number): Promise<void> {
+  public async deleteOne(id: number, userId: number): Promise<void> {
     await this.repository
       .createQueryBuilder('mealTicketRemoval')
       .delete()
       .where('mealTicketRemoval.id = :id', { id })
+      .andWhere('mealTicketRemoval.userId = :userId', { userId })
       .execute();
   }
 }
